fix: use base class instead of the interface
