+++ doc/10_Extending_Studio/05_Perspectives/01_Widgets.md
@@ -53,20 +53,20 @@ declare(strict_types=1);
 namespace App\Perspective\Widget\Hydrator;
 
 use App\Perspective\Widget\Model\CustomWidgetConfig;
-use Pimcore\Bundle\StudioBackendBundle\Perspective\Model\WidgetConfigInterface;
 
 final readonly class CustomWidgetConfigHydrator implements WidgetConfigHydratorInterface
 {
     public function getSupportedWidgetType(): string
     {
-        return 'my_custom_widget';
+        return 'my_custom_widget_type';
     }
 
-    public function hydrate(array $widgetData): WidgetConfigInterface
+    public function hydrate(array $widgetData): CustomWidgetConfig
     {
         return new CustomWidgetConfig(
             $widgetData['id'],
             $widgetData['name'],
+            $widgetData['icon'],
             $widgetData['customField'],
             $widgetData['createdAt'],
             $widgetData['isWriteable']
@@ -75,9 +75,62 @@ final readonly class CustomWidgetConfigHydrator implements WidgetConfigHydratorI
 }
 ```
 
+### Example Widget Configuration Schema
+
+```php
+<?php
+declare(strict_types=1);
+
+namespace App\Perspective\Schema;
+
+use App\Perspective\Widget\Model\CustomWidgetConfig;
+use Pimcore\Bundle\StudioBackendBundle\Perspective\Schema\WidgetConfig;
+
+#[Schema(
+    title: 'Custom Widget',
+    required: [
+        'customField',
+        'createdAt',
+        'isWriteable',
+    ],
+    type: 'object'
+)]
+final class MyCustomWidget extends WidgetConfig
+{
+    public function __construct(
+        string $id,
+        string $name,
+        ?ElementIcon $icon = null,
+        #[Property(description: 'Custom Field', type: 'string', example: 'Some Value')]
+        private readonly string $customField = 'Some Value',
+        #[Property(description: 'Created At timestamp', type: 'int', example: 1738917191)]
+        private readonly ?int $createdAt = null,
+        #[Property(description: 'Is Writeable', type: 'bool', example: false)]
+        private readonly bool $isWriteable = false,
+    ) {
+        parent::__construct($id, $name, 'my_custom_widget_type', $icon);
+    }
+    
+    public function getCustomField(): string
+    {
+        return $this->customField;
+    }
+    
+    public function getCreatedAt(): ?int
+    {
+        return $this->createdAt;
+    }
+    
+    public function isWriteable(): bool
+    {
+        return $this->isWriteable;
+    }
+}
+```
+
 :::info
 
-Please note that the Hydrator must return a class implementing `WidgetConfigInterface` marker interface.
+Please note that the Hydrator must return a class extending the `Pimcore\Bundle\StudioBackendBundle\Perspective\Schema\WidgetConfig` class.
 
 :::
